+++ include/model.h
@@ -2,13 +2,15 @@
 #define MODEL_H
 
 #include <stdint.h>
+
 #include "cglm/struct.h"
 
 #include "mesh.h"
+#include "core/defines.h"
 
 struct model {
         struct mesh **mesh;
-        uint32_t mesh_count;
+        u32 mesh_count;
 
         mat4s model;
         mat4s view;
@@ -21,7 +23,7 @@ struct model *model_create();
  * we load other models as well, we might want to either
  * pass a model type in this function, or create multiple functions
  * to load different types of models.*/
-int model_load_from_file(struct model *model, const char *filepath);
+i8 model_load_from_file(struct model *model, const char *filepath);
 void model_destroy(struct model *model);
 
 #endif
